tests: Move stuff to test_dimension.py, better docs
