+++ tests/test_dimension.py
@@ -2,89 +2,168 @@ import numpy as np
 
 from conftest import skipif_yask
 
-from devito import ConditionalDimension, Grid, TimeFunction, Eq, Operator, Constant
+from devito import (ConditionalDimension, Grid, TimeFunction, Eq, Operator, Constant,
+                    INTERIOR)
+from devito.ir.iet import retrieve_iteration_tree
 
 
 @skipif_yask
-def test_conditional_basic():
-    nt = 19
-    grid = Grid(shape=(11, 11))
-    time = grid.time_dim
+class TestSubDimension(object):
 
-    u = TimeFunction(name='u', grid=grid)
-    assert(grid.stepping_dim in u.indices)
+    def test_domain_vs_interior(self):
+        """
+        Tests regions work properly in terms of code generation and runtime
+        argument derivation. All regions but the default one, ``DOMAIN``,
+        induce one or more :class:`SubDimension`s.
+        """
+        grid = Grid(shape=(4, 4, 4))
+        x, y, z = grid.dimensions
+        t = grid.stepping_dim  # noqa
 
-    u2 = TimeFunction(name='u2', grid=grid, save=nt)
-    assert(time in u2.indices)
+        u = TimeFunction(name='u', grid=grid)  # noqa
+        eqs = [Eq(u.forward, u + 1),
+               Eq(u.forward, u.forward + 2, region=INTERIOR)]
 
-    factor = 4
-    time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
-    usave = TimeFunction(name='usave', grid=grid, save=(nt+factor-1)//factor,
-                         time_dim=time_subsampled)
-    assert(time_subsampled in usave.indices)
+        op = Operator(eqs, dse='noop', dle='noop')
+        trees = retrieve_iteration_tree(op)
+        assert len(trees) == 2
 
-    eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.), Eq(usave, u)]
-    op = Operator(eqns)
-    op.apply(t_M=nt-2)
-    assert np.all(np.allclose(u.data[(nt-1) % 3], nt-1))
-    assert np.all([np.allclose(u2.data[i], i) for i in range(nt)])
-    assert np.all([np.allclose(usave.data[i], i*factor)
-                  for i in range((nt+factor-1)//factor)])
+        op.apply(time_M=1)
+        assert np.all(u.data[1, 0, :, :] == 1)
+        assert np.all(u.data[1, -1, :, :] == 1)
+        assert np.all(u.data[1, :, 0, :] == 1)
+        assert np.all(u.data[1, :, -1, :] == 1)
+        assert np.all(u.data[1, :, :, 0] == 1)
+        assert np.all(u.data[1, :, :, -1] == 1)
+        assert np.all(u.data[1, 1:3, 1:3, 1:3] == 3)
+
+    def test_flow_detection_interior(self):
+        """
+        Test detection of flow directions when :class:`SubDimension`s are used
+        (in this test they are induced by the ``INTERIOR`` region).
+
+        Stencil uses values at new timestep as well as those at previous ones
+        This forces an evaluation order onto x.
+        Weights are:
+
+               x=0     x=1     x=2     x=3
+         t=N    2    ---3
+                v   /
+         t=N+1  o--+----4
+
+        Flow dependency should traverse x in the negative direction
+
+               x=2     x=3     x=4     x=5      x=6
+        t=0             0   --- 0     -- 1    -- 0
+                        v  /    v    /   v   /
+        t=1            44 -+--- 11 -+--- 2--+ -- 0
+        """
+
+        grid = Grid(shape=(10, 10))
+        x, y = grid.dimensions
+        u = TimeFunction(name='u', grid=grid, save=10, time_order=1, space_order=0)
+        step = Eq(u.forward, 2*u
+                  + 3*u.subs(x, x+x.spacing)
+                  + 4*u.forward.subs(x, x+x.spacing),
+                  region=INTERIOR)
+        op = Operator(step)
+
+        u.data[0, 5, 5] = 1.0
+        op.apply(time_M=0)
+        assert u.data[1, 5, 5] == 2
+        assert u.data[1, 4, 5] == 11
+        assert u.data[1, 3, 5] == 44
+        assert u.data[1, 2, 5] == 4*44
+        assert u.data[1, 1, 5] == 4*4*44
+
+        # This point isn't updated because of the INTERIOR selection
+        assert u.data[1, 0, 5] == 0
+
+        assert np.all(u.data[1, 6:, :] == 0)
+        assert np.all(u.data[1, :, 0:5] == 0)
+        assert np.all(u.data[1, :, 6:] == 0)
 
 
-@skipif_yask
-def test_conditional_as_expr():
-    nt = 19
-    grid = Grid(shape=(11, 11))
-    time = grid.time_dim
+class TestConditionalDimension(object):
 
-    u = TimeFunction(name='u', grid=grid)
-    assert(grid.stepping_dim in u.indices)
+    @skipif_yask
+    def test_basic(self):
+        nt = 19
+        grid = Grid(shape=(11, 11))
+        time = grid.time_dim
 
-    u2 = TimeFunction(name='u2', grid=grid, save=nt)
-    assert(time in u2.indices)
+        u = TimeFunction(name='u', grid=grid)
+        assert(grid.stepping_dim in u.indices)
 
-    factor = 4
-    time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
-    usave = TimeFunction(name='usave', grid=grid, save=(nt+factor-1)//factor,
-                         time_dim=time_subsampled)
-    assert(time_subsampled in usave.indices)
+        u2 = TimeFunction(name='u2', grid=grid, save=nt)
+        assert(time in u2.indices)
 
-    eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.),
-            Eq(usave, time_subsampled * u)]
-    op = Operator(eqns)
-    op.apply(t=nt-2)
-    assert np.all(np.allclose(u.data[(nt-1) % 3], nt-1))
-    assert np.all([np.allclose(u2.data[i], i) for i in range(nt)])
-    assert np.all([np.allclose(usave.data[i], i*factor*i)
-                  for i in range((nt+factor-1)//factor)])
+        factor = 4
+        time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
+        usave = TimeFunction(name='usave', grid=grid, save=(nt+factor-1)//factor,
+                             time_dim=time_subsampled)
+        assert(time_subsampled in usave.indices)
 
+        eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.), Eq(usave, u)]
+        op = Operator(eqns)
+        op.apply(t_M=nt-2)
+        assert np.all(np.allclose(u.data[(nt-1) % 3], nt-1))
+        assert np.all([np.allclose(u2.data[i], i) for i in range(nt)])
+        assert np.all([np.allclose(usave.data[i], i*factor)
+                      for i in range((nt+factor-1)//factor)])
 
-@skipif_yask
-def test_conditional_shifted():
-    nt = 19
-    grid = Grid(shape=(11, 11))
-    time = grid.time_dim
+    @skipif_yask
+    def test_as_expr(self):
+        nt = 19
+        grid = Grid(shape=(11, 11))
+        time = grid.time_dim
 
-    u = TimeFunction(name='u', grid=grid)
-    assert(grid.stepping_dim in u.indices)
+        u = TimeFunction(name='u', grid=grid)
+        assert(grid.stepping_dim in u.indices)
 
-    u2 = TimeFunction(name='u2', grid=grid, save=nt)
-    assert(time in u2.indices)
+        u2 = TimeFunction(name='u2', grid=grid, save=nt)
+        assert(time in u2.indices)
 
-    factor = 4
-    time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
-    usave = TimeFunction(name='usave', grid=grid, save=2, time_dim=time_subsampled)
-    assert(time_subsampled in usave.indices)
+        factor = 4
+        time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
+        usave = TimeFunction(name='usave', grid=grid, save=(nt+factor-1)//factor,
+                             time_dim=time_subsampled)
+        assert(time_subsampled in usave.indices)
 
-    t_sub_shift = Constant(name='t_sub_shift', dtype=np.int32)
+        eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.),
+                Eq(usave, time_subsampled * u)]
+        op = Operator(eqns)
+        op.apply(t=nt-2)
+        assert np.all(np.allclose(u.data[(nt-1) % 3], nt-1))
+        assert np.all([np.allclose(u2.data[i], i) for i in range(nt)])
+        assert np.all([np.allclose(usave.data[i], i*factor*i)
+                      for i in range((nt+factor-1)//factor)])
 
-    eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.),
-            Eq(usave.subs(time_subsampled, time_subsampled - t_sub_shift), u)]
-    op = Operator(eqns)
+    @skipif_yask
+    def test_shifted(self):
+        nt = 19
+        grid = Grid(shape=(11, 11))
+        time = grid.time_dim
 
-    # Starting at time_m=10, so time_subsampled - t_sub_shift is in range
-    op.apply(time_m=10, time_M=nt-2, t_sub_shift=3)
-    assert np.all(np.allclose(u.data[0], 8))
-    assert np.all([np.allclose(u2.data[i], i - 10) for i in range(10, nt)])
-    assert np.all([np.allclose(usave.data[i], 2+i*factor) for i in range(2)])
+        u = TimeFunction(name='u', grid=grid)
+        assert(grid.stepping_dim in u.indices)
+
+        u2 = TimeFunction(name='u2', grid=grid, save=nt)
+        assert(time in u2.indices)
+
+        factor = 4
+        time_subsampled = ConditionalDimension('t_sub', parent=time, factor=factor)
+        usave = TimeFunction(name='usave', grid=grid, save=2, time_dim=time_subsampled)
+        assert(time_subsampled in usave.indices)
+
+        t_sub_shift = Constant(name='t_sub_shift', dtype=np.int32)
+
+        eqns = [Eq(u.forward, u + 1.), Eq(u2.forward, u2 + 1.),
+                Eq(usave.subs(time_subsampled, time_subsampled - t_sub_shift), u)]
+        op = Operator(eqns)
+
+        # Starting at time_m=10, so time_subsampled - t_sub_shift is in range
+        op.apply(time_m=10, time_M=nt-2, t_sub_shift=3)
+        assert np.all(np.allclose(u.data[0], 8))
+        assert np.all([np.allclose(u2.data[i], i - 10) for i in range(10, nt)])
+        assert np.all([np.allclose(usave.data[i], 2+i*factor) for i in range(2)])
